pyproject.toml
[tool.poetry]
name = "onlineplatform"
version = "0.1.0"
description = ""
authors = ["pavellis <manbyak5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "2.9.4"
django-filter = "^22.1"
django-environ = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
